requirements.txt
annotated-types==0.6.0
ansicon==1.89.0
anyio==4.3.0
ascii-magic==2.3.0
asgiref==3.8.1
blessed==1.20.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
dj-database-url==2.2.0
Django==5.0.3
django-cors-headers==4.3.1
django-heroku==0.3.1
djangorestframework==3.15.2
filelock==3.13.4
fsspec==2024.3.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
Jinja2==3.1.3
jinxed==1.2.1
keyboard==0.13.5
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
nodejs-bin==18.4.0a4
numpy==1.26.4
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
SpeechRecognition==3.10.4
sqlparse==0.4.4
sympy==1.12
tokenizers==0.15.2
tqdm==4.66.2
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
wheel==0.43.0